+++ app/src/assets/css/Application.css
@@ -2,6 +2,11 @@
   margin-top: 1rem;
 }
 
+.alternative {
+  margin-top: 1rem;
+  background: #ffffff;
+}
+
 .content {
   display: flex;
   flex-wrap: wrap;
@@ -9,4 +14,8 @@
   align-items: center;
   max-width: 64rem;
   margin: 0 auto;
+}
+
+.nav {
+  padding-top: 1rem;
 }
